# Conflicts:
#	src/main/java/com/example/telegramBot/service/SendBotMessageService.java
#	src/main/java/com/example/telegramBot/service/SendBotMessageServiceRealisation.java
#	src/main/java/com/example/telegramBot/student/command/CommandName.java
#	src/main/java/com/example/telegramBot/student/command/StudentCommandHandler.java
#	src/main/java/com/example/telegramBot/student/command/commands/UnknownComm.java
#	src/main/java/com/example/telegramBot/user/keyboard/inline/UserInlineKeyboardSource.java

+++ src/main/java/com/example/telegramBot/user/command/commands/TestComm.java
@@ -2,6 +2,7 @@ package com.example.telegramBot.user.command.commands;
 
 import com.example.telegramBot.service.SendBotMessageService;
 import com.example.telegramBot.user.keyboard.inline.UserInlineKeyboardSource;
+import com.example.telegramBot.user.keyboard.reply.UserReplyKeyboardSource;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboard;
 
@@ -18,6 +19,9 @@ public class TestComm implements Command {
 
     UserInlineKeyboardSource userInlineKeyboardSource = new UserInlineKeyboardSource();
 
+    private final UserReplyKeyboardSource userReplyKeyboardSource = new UserReplyKeyboardSource();
+    public final ReplyKeyboard returnToMainMenu = userReplyKeyboardSource.getReturnToMainMenu();
+
     public TestComm(SendBotMessageService sendBotMessageService) {
         this.sendBotMessageService = sendBotMessageService;
     }
@@ -28,7 +32,10 @@ public class TestComm implements Command {
         String chatId = update.getCallbackQuery().getMessage().getChatId().toString();
         Integer message_id = update.getCallbackQuery().getMessage().getMessageId();
 
-        sendBotMessageService.deleteMessage(chatId, message_id);
+        if (numberOfQuestion > 0) {
+            sendBotMessageService.deleteMessage(chatId, message_id);
+        }
+
 
         addTheAnswerIntoList(update);
 
@@ -55,11 +62,12 @@ public class TestComm implements Command {
             }
         }
         int result = count * 100 / TEST_QUESTIONS.size();
-        sendBotMessageService.sendMessage(chatId, String.format("Тест пройден. Ваш результат %d %%", result));
+        sendBotMessageService.sendMessage(chatId, String.format("Тест пройден. Ваш результат %d %%", result), returnToMainMenu);
+
     }
 
     private void addTheAnswerIntoList(Update update) {
-        if (update.getCallbackQuery().getData().startsWith("english")) {
+        if (!update.getCallbackQuery().getData().equals("english")) {
             answers.add(update.getCallbackQuery().getData().replace("english", ""));
         }
     }
